requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# AI and Processing
openai==1.3.7
pandas==2.1.4
nltk==3.8.1
spacy==3.7.2

# External Integrations
microsoft-graph-auth==0.1.4
atlassian-python-api==3.41.9
exchangelib==5.0.3
requests==2.31.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
croniter==1.4.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8 